vitamin-app-frontend/vitamin-backend/requirements.txt
fastapi
uvicorn
motor==3.3.2
pymongo==4.6.3
beanie==1.25.0
pydantic==2.6.1
python-dotenv
httpx
passlib[bcrypt]
python-multipart
python-jose[cryptography]